Backend

+++ routes/notes.js
@@ -1,24 +1,32 @@
 const express= require('express')
+const auth= require('../src/middleware/auth')
 const Note=require('../src/models/notemodel')
 //const router=express.Router()
 const app=express()
 app.use(express.json())
-app.get('/notes',(req,res)=>{
-    Note.find({}).then((Note)=>{
-   res.send(Note);
-    }).catch((error)=>{
+app.get('/notes',auth,async (req,res)=>{
+    try{
+    await req.user.populate('notes').execPopulate()
+
+    res.send(req.user.notes)
+
+    }
+    
+    catch(e){
    res.render(error)
-    })
+    }
   
 })
 
-app.post('/notes',(req,res)=>{
+app.post('/notes',auth ,(req,res)=>{
     const reqbody=req.body
     console.log('reqbody')
     console.log(req.body)
     const note1= new Note(
-        
-         reqbody   
+        {
+         ...reqbody,
+         owner: req.user._id   
+        }
         )
         note1.save().then(()=>{
             console.log(note1)
@@ -37,14 +45,16 @@ app.get('/notes/:id',(req,res)=>{
         res.send(note)
     }).catch((error)=>{res.send(error)})
 })
-
-app.delete('/notes/delete/:id',(req,res)=>{
+//delete notes
+app.delete('/notes/delete/:id',auth,(req,res)=>{  
 const _id=req.params.id;
-console.log(_id)
-res.send('success')
-Note.deleteOne({_id:_id}).then((note)=>{
+try{
+Note.deleteOne({_id:_id}, {owner:req.user._id}).then((note)=>{
     res.status(200).send('Deleted')
-}).catch((error)=>{res.send(error)})
-
+})
+}
+catch(e){
+  res.status(404).send()  
+}
 })
 module.exports=app
